# even faster with Cython!