--- a/pi_estimate_in_C/run_me.sh
+++ b/pi_estimate_in_C/run_me.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+g++ -O2 -o pi_estimate pi_estimate.cpp &&
+./pi_estimate