--- a/single_test_run.py
+++ b/single_test_run.py
@@ -11,6 +11,7 @@ import random
 import time
 import argparse
 
+from numba import njit
 
 def estimate_pi(
     n_points: int,
@@ -41,11 +42,44 @@ def estimate_pi(
     if not show_estimate:
         print("Final Estimation of Pi=", pi_estimate)
 
+@njit
+def estimate_pi_numba(
+    n_points: int,
+    show_estimate: bool,
+) -> None:
+    """
+    Simple Monte Carlo Pi estimation calculation.
+
+    Parameters
+    ----------
+    n_points
+        number of random numbers used to for estimation.
+    show_estimate
+        if True, will show the estimation of Pi, otherwise
+        will not output anything.
+    """
+    within_circle = 0
+
+    for _ in range(n_points):
+        x = random.uniform(-1, 1)
+        y = random.uniform(-1, 1)
+        radius_squared = x**2 + y**2
+
+        if radius_squared <= 1:
+            within_circle += 1
+
+    pi_estimate = 4 * within_circle / n_points
+
+    if not show_estimate:
+        print("Final Estimation of Pi=", pi_estimate)
+
+
 
 def run_test(
     n_points: int,
     n_repeats: int,
     only_time: bool,
+    jit: str
 ) -> None:
     """
     Perform the tests and measure required time.
@@ -60,11 +94,17 @@ def run_test(
         if True will only print the time, otherwise
         will also show the Pi estimate and a neat formatted
         time.
+    jit
+        type of jit to use none,numba
     """
+    fcns = {"none":estimate_pi, "numba": estimate_pi_numba}
+
     start_time = time.time()
 
+
+
     for _ in range(n_repeats):
-        estimate_pi(n_points, only_time)
+        fcns[jit](n_points, only_time)
 
     if only_time:
         print(f"{(time.time() - start_time)/n_repeats:.4f}")
@@ -95,6 +135,14 @@ def main(arguments=None):
         type=positive_integer,
         default=1_000_000,
     )
+
+    parser.add_argument(
+        "--jit",
+        help="use compiler options=[none, numba]",
+        default="none",
+        type=str
+    )
+
     parser.add_argument(
         "-r",
         "--n_repeats",
@@ -109,10 +157,12 @@ def main(arguments=None):
     )
     args = parser.parse_args()
 
+
     run_test(
         args.n_points,
         args.n_repeats,
         args.only_time,
+        args.jit
     )